refactor: move fact filtering to FilteredPool impl. and introduce FilteredPool trait
